+++ CHANGES.md
@@ -9,6 +9,7 @@ the one you find pypi, but its the one in development.
 * native support for stop_cmd, (pre|post)_start_cmd and post_stop_cmd in pystemd.run.
 * adding lxml as a dependency.
 * drop support for python 3.4 and 3.5 (yey welcome f-strings)
+* change using select.select to use DefaultSelector in pystemd.run
 
 0.10.0
 =====